requirements.txt
python-telegram-bot==20.7
geopy==2.4.1
python-dotenv==1.0.0
pandas==2.0.3
python-dateutil==2.8.2